SecurityAnalysisResultBuilder, modify API to make it inheritable in other projects
